requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.0.1
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
ecdsa==0.19.0
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
passlib==1.7.4
playwright==1.50.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
PyJWT==2.10.1
python-jose==3.4.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.46.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
